requirements.txt
google-genai
python-dotenv
streamlit
Pillow
openpyxl
